+++ adapters/repos/db/vector/geo/coordinates_for_id.go
@@ -20,12 +20,12 @@ import (
 
 // CoordinatesForID must provide the geo coordinates for the specified index
 // id
-type CoordinatesForID func(ctx context.Context, id int32) (*models.GeoCoordinates, error)
+type CoordinatesForID func(ctx context.Context, id uint64) (*models.GeoCoordinates, error)
 
 // VectorForID transforms the geo coordinates into a "vector" of fixed length
 // two, where element 0 represents the latitude and element 1 represents the
 // longitude. This way it is usable by a generic vector index such as HNSW
-func (cfid CoordinatesForID) VectorForID(ctx context.Context, id int32) ([]float32, error) {
+func (cfid CoordinatesForID) VectorForID(ctx context.Context, id uint64) ([]float32, error) {
 	coordinates, err := cfid(ctx, id)
 	if err != nil {
 		return nil, err
