+++ adapters/repos/db/vector/hnsw/search.go
@@ -32,23 +32,22 @@ func reasonableEfFromK(k int) int {
 	return ef
 }
 
-func (h *hnsw) SearchByID(id int, k int) ([]int, error) {
+func (h *hnsw) SearchByID(id uint64, k int) ([]uint64, error) {
 	return h.knnSearch(id, k, reasonableEfFromK(k))
 }
 
-func (h *hnsw) SearchByVector(vector []float32, k int, allowList helpers.AllowList) ([]int, error) {
+func (h *hnsw) SearchByVector(vector []float32, k int, allowList helpers.AllowList) ([]uint64, error) {
 	return h.knnSearchByVector(vector, k, reasonableEfFromK(k), allowList)
 }
 
-func (h *hnsw) knnSearch(queryNodeID int, k int, ef int) ([]int, error) {
+func (h *hnsw) knnSearch(queryNodeID uint64, k int, ef int) ([]uint64, error) {
 	entryPointID := h.entryPointID
 	entryPointDistance, ok, err := h.distBetweenNodes(entryPointID, queryNodeID)
 	if err != nil || !ok {
 		return nil, errors.Wrap(err, "knn search: distance between entrypint and query node")
 	}
 
-	queryVector, err := h.vectorForID(context.Background(),
-		int32(queryNodeID))
+	queryVector, err := h.vectorForID(context.Background(), queryNodeID)
 	if err != nil {
 		return nil, errors.Wrapf(err, "could not get vector of object at docID %d", queryNodeID)
 	}
@@ -75,7 +74,7 @@ func (h *hnsw) knnSearch(queryNodeID int, k int, ef int) ([]int, error) {
 
 	flat := res.flattenInOrder()
 	size := min(len(flat), k)
-	out := make([]int, size)
+	out := make([]uint64, size)
 	for i, elem := range flat {
 		if i >= size {
 			break
@@ -106,7 +105,7 @@ func (h *hnsw) searchLayerByVector(queryVector []float32,
 			return nil, errors.Wrapf(err, "calculate distance of current last result")
 		}
 
-		dist, ok, err := h.distanceToNode(distancer, int32(candidate.index))
+		dist, ok, err := h.distanceToNode(distancer, candidate.index)
 		if err != nil {
 			return nil, errors.Wrap(err, "calculate distance between candidate and query")
 		}
@@ -147,10 +146,10 @@ func (h *hnsw) searchLayerByVector(queryVector []float32,
 	return results, nil
 }
 
-func newVisitedList(entrypoints binarySearchTreeGeneric) map[uint32]struct{} {
-	visited := map[uint32]struct{}{}
+func newVisitedList(entrypoints binarySearchTreeGeneric) map[uint64]struct{} {
+	visited := map[uint64]struct{}{}
 	for _, elem := range entrypoints.flattenInOrder() {
-		visited[uint32(elem.index)] = struct{}{}
+		visited[elem.index] = struct{}{}
 	}
 	return visited
 }
@@ -165,7 +164,7 @@ func (h *hnsw) insertViableEntrypointsAsCandidatesAndResults(
 			// have an allow list (i.e. the user has probably set some sort of a
 			// filter restricting this search further. As a result we have to
 			// ignore items not on the list
-			if !allowList.Contains(uint32(ep.index)) {
+			if !allowList.Contains(ep.index) {
 				continue
 			}
 		}
@@ -181,7 +180,7 @@ func (h *hnsw) insertViableEntrypointsAsCandidatesAndResults(
 func (h *hnsw) currentWorstResultDistance(results *binarySearchTreeGeneric,
 	distancer distancer.Distancer) (float32, error) {
 	if results.root != nil {
-		id := int32(results.maximum().index)
+		id := results.maximum().index
 		d, ok, err := h.distanceToNode(distancer, id)
 		if err != nil {
 			return 0, errors.Wrap(err,
@@ -202,8 +201,8 @@ func (h *hnsw) currentWorstResultDistance(results *binarySearchTreeGeneric,
 }
 
 func (h *hnsw) extendCandidatesAndResultsFromNeighbors(candidates,
-	results *binarySearchTreeGeneric, connections []uint32,
-	visited map[uint32]struct{}, distancer distancer.Distancer, ef int,
+	results *binarySearchTreeGeneric, connections []uint64,
+	visited map[uint64]struct{}, distancer distancer.Distancer, ef int,
 	level int, allowList helpers.AllowList, worstResultDistance float32) error {
 	for _, neighborID := range connections {
 		if _, ok := visited[neighborID]; ok {
@@ -214,7 +213,7 @@ func (h *hnsw) extendCandidatesAndResultsFromNeighbors(candidates,
 		// make sure we never visit this neighbor again
 		visited[neighborID] = struct{}{}
 
-		distance, ok, err := h.distanceToNode(distancer, int32(neighborID))
+		distance, ok, err := h.distanceToNode(distancer, neighborID)
 		if err != nil {
 			return errors.Wrap(err, "calculate distance between candidate and query")
 		}
@@ -226,7 +225,7 @@ func (h *hnsw) extendCandidatesAndResultsFromNeighbors(candidates,
 
 		resLenBefore := results.len() // calculating just once saves a bit of time
 		if distance < worstResultDistance || resLenBefore < ef {
-			candidates.insert(int(neighborID), distance)
+			candidates.insert(neighborID, distance)
 			if level == 0 && allowList != nil {
 				// we are on the lowest level containing the actual candidates and we
 				// have an allow list (i.e. the user has probably set some sort of a
@@ -237,11 +236,11 @@ func (h *hnsw) extendCandidatesAndResultsFromNeighbors(candidates,
 				}
 			}
 
-			if h.hasTombstone(int(neighborID)) {
+			if h.hasTombstone(neighborID) {
 				continue
 			}
 
-			results.insert(int(neighborID), distance)
+			results.insert(neighborID, distance)
 
 			// +1 because we have added one node size calculating the len
 			if resLenBefore+1 > ef {
@@ -255,7 +254,7 @@ func (h *hnsw) extendCandidatesAndResultsFromNeighbors(candidates,
 }
 
 func (h *hnsw) distanceToNode(distancer distancer.Distancer,
-	nodeID int32) (float32, bool, error) {
+	nodeID uint64) (float32, bool, error) {
 	candidateVec, err := h.vectorForID(context.Background(), nodeID)
 	if err != nil {
 		var e storobj.ErrNotFound
@@ -279,14 +278,14 @@ func (h *hnsw) distanceToNode(distancer distancer.Distancer,
 // the underlying object seems to have been deleted, to recover from
 // this situation let's add a tombstone to the deleted object, so it
 // will be cleaned up and skip this candidate in the current search
-func (h *hnsw) handleDeletedNode(docID int32) {
-	if h.hasTombstone(int(docID)) {
+func (h *hnsw) handleDeletedNode(docID uint64) {
+	if h.hasTombstone(docID) {
 		// nothing to do, this node already has a tombstone, it will be cleaned up
 		// in the next deletion cycle
 		return
 	}
 
-	h.addTombstone(int(docID))
+	h.addTombstone(docID)
 	h.logger.WithField("action", "attach_tombstone_to_deleted_node").
 		WithField("node_id", docID).
 		Info("found a deleted node (%d) without a tombstone, "+
@@ -294,7 +293,7 @@ func (h *hnsw) handleDeletedNode(docID int32) {
 }
 
 func (h *hnsw) knnSearchByVector(searchVec []float32, k int,
-	ef int, allowList helpers.AllowList) ([]int, error) {
+	ef int, allowList helpers.AllowList) ([]uint64, error) {
 	if h.isEmpty() {
 		return nil, nil
 	}
@@ -333,7 +332,7 @@ func (h *hnsw) knnSearchByVector(searchVec []float32, k int,
 
 	flat := res.flattenInOrder()
 	size := min(len(flat), k)
-	out := make([]int, size)
+	out := make([]uint64, size)
 	for i, elem := range flat {
 		if i >= size {
 			break
@@ -345,32 +344,32 @@ func (h *hnsw) knnSearchByVector(searchVec []float32, k int,
 }
 
 func (h *hnsw) selectNeighborsSimple(input binarySearchTreeGeneric,
-	max int, denyList helpers.AllowList) []uint32 {
+	max int, denyList helpers.AllowList) []uint64 {
 	flat := input.flattenInOrder()
 
 	maxSize := min(len(flat), max)
-	out := make([]uint32, maxSize)
+	out := make([]uint64, maxSize)
 	actualSize := 0
 	for i, elem := range flat {
-		if denyList != nil && denyList.Contains(uint32(elem.index)) {
+		if denyList != nil && denyList.Contains(elem.index) {
 			continue
 		}
 
 		if i >= maxSize {
 			break
 		}
-		out[actualSize] = uint32(elem.index)
+		out[actualSize] = elem.index
 		actualSize++
 	}
 
 	return out[:actualSize]
 }
 
-func (h *hnsw) selectNeighborsSimpleFromId(nodeId int, ids []uint32,
-	max int, denyList helpers.AllowList) ([]uint32, error) {
+func (h *hnsw) selectNeighborsSimpleFromId(nodeId uint64, ids []uint64,
+	max int, denyList helpers.AllowList) ([]uint64, error) {
 	bst := &binarySearchTreeGeneric{}
 	for _, id := range ids {
-		dist, ok, err := h.distBetweenNodes(int(id), nodeId)
+		dist, ok, err := h.distBetweenNodes(id, nodeId)
 		if err != nil {
 			return nil, errors.Wrap(err, "select neighbors simple from id")
 		}
@@ -379,7 +378,7 @@ func (h *hnsw) selectNeighborsSimpleFromId(nodeId int, ids []uint32,
 			// node was deleted in the underlying object store
 			continue
 		}
-		bst.insert(int(id), dist)
+		bst.insert(id, dist)
 	}
 
 	return h.selectNeighborsSimple(*bst, max, denyList), nil
